Move TChar from TBuffer.cpp to a separate file
