+++ src/CMakeLists.txt
@@ -76,6 +76,7 @@ set(mudlet_SRCS
     TAlias.cpp
     TArea.cpp
     TBuffer.cpp
+    TChar.cpp
     TCommandLine.cpp
     TConsole.cpp
     TDebug.cpp
@@ -203,6 +204,7 @@ set(mudlet_HDRS
     TArea.h
     TAstar.h
     TBuffer.h
+    TChar.h
     TCommandLine.h
     TConsole.h
     TDebug.h
