+++ src/TBuffer.cpp
@@ -26,6 +26,7 @@
 #include "TConsole.h"
 #include "TStringUtils.h"
 
+#include "TChar.h"
 #include "pre_guard.h"
 #include <QTextBoundaryFinder>
 #include <QTextCodec>
@@ -50,71 +51,6 @@
 // of the type (?) of an `\x1b[?z` received}:
 //#define DEBUG_MXP_PROCESSING
 
-
-// Default constructor:
-TChar::TChar()
-: mFgColor(Qt::white)
-, mBgColor(Qt::black)
-, mFlags(None)
-, mIsSelected(false)
-, mLinkIndex(0)
-{
-}
-
-TChar::TChar(const QColor& fg, const QColor& bg, const TChar::AttributeFlags flags, const int linkIndex)
-: mFgColor(fg)
-, mBgColor(bg)
-, mFlags(flags)
-, mIsSelected(false)
-, mLinkIndex(linkIndex)
-{
-}
-
-TChar::TChar(Host* pH)
-: mFlags(None)
-, mIsSelected(false)
-, mLinkIndex(0)
-{
-    if (pH) {
-        mFgColor = pH->mFgColor;
-        mBgColor = pH->mBgColor;
-    } else {
-        mFgColor = Qt::white;
-        mBgColor = Qt::black;
-    }
-}
-
-// Note: this operator compares ALL aspects of 'this' against 'other' which may
-// not be wanted in every case:
-bool TChar::operator==(const TChar& other)
-{
-    if (mIsSelected != other.mIsSelected) {
-        return false;
-    }
-    if (mLinkIndex != other.mLinkIndex) {
-        return false;
-    }
-    if (mFgColor != other.mFgColor) {
-        return false;
-    }
-    if (mBgColor != other.mBgColor) {
-        return false;
-    }
-    if (mFlags != other.mFlags) {
-        return false;
-    }
-    return true;
-}
-
-// Copy constructor:
-TChar::TChar(const TChar& copy)
-: mFgColor(copy.mFgColor)
-, mBgColor(copy.mBgColor)
-, mFlags(copy.mFlags)
-, mIsSelected(false)
-, mLinkIndex(copy.mLinkIndex)
-{
-}
 
 const QString timeStampFormat = QStringLiteral("hh:mm:ss.zzz ");
 const QString blankTimeStamp  = QStringLiteral("------------ ");
