+++ src/TConsole.h
@@ -30,11 +30,12 @@
 
 #include "TTextCodec.h"
 
+#include "TChar.h"
 #include "pre_guard.h"
+#include <QTextStream>
 #include <QDataStream>
 #include <QFile>
 #include <QPointer>
-#include <QTextStream>
 #include <QWidget>
 #include "post_guard.h"
