+++ src/TTextEdit.h
@@ -28,13 +28,14 @@
 
 #include "TBuffer.h"
 
+#include "TChar.h"
 #include "pre_guard.h"
+#include <QTime>
 #include <QMap>
 #include <QPointer>
-#include <QTime>
 #include <QWidget>
-#include <chrono>
 #include "post_guard.h"
+#include <chrono>
 
 #include <string>
